Remove ContactsData from CommonData
This is because ContactsData comes from a different data source
(that may not be available) than the rest of the topology objects.
Now I have to pass ContactsData down to the methods that need it;
this is annoying but it lets me save the other topology objects
without embedding the ContactsData (or lack thereof) into them.